requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Django==4.0.6
django-choices==1.7.2
djangorestframework==3.13.1
greenlet==1.1.2
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.4.0
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.1
requests==2.28.1
sendgrid==6.9.7
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tomli==2.0.1
twilio==7.10.0
typing-extensions==4.3.0
urllib3==1.26.10